+++ ingest/operation_details.go
@@ -362,6 +362,10 @@ func (is *Session) operationDetails() map[string]interface{} {
 			details["poll_id"] = op.Data.MustRemoveData().PollId
 		}
 	case xdr.OperationTypeManageAccountSpecificRule:
+	case xdr.OperationTypeRemoveAssetPair:
+		op := c.Operation().Body.MustRemoveAssetPairOp()
+		details["base"] = op.Base
+		details["quote"] = op.Quote
 	case xdr.OperationTypeCreateKycRecoveryRequest:
 		op := c.Operation().Body.MustCreateKycRecoveryRequestOp()
 		details["target_account"] = op.TargetAccount